Updated import alias, moved to navigation folder
