+++ src/components/navigation/SearchBox.jsx
@@ -1,8 +1,8 @@
 
 'use-client'
 
-import products from "../DB/product.json"
-import { usePathname, useRouter } from 'next/navigation';
+import products from "@/DB/product.json"
+import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react'
 import { RiSearchLine } from "react-icons/ri"
 const SearchBox = () => {
@@ -35,18 +35,18 @@ const SearchBox = () => {
 
     return (
         <>
-            <form className="relative w-36 xs:w-48 sm:w-72 h-6 xs:h-8 sm:h-9 md:h-11 " onSubmit={handleSearch}>
+            <form className="relative h-6 w-36 xs:w-48 sm:w-72 xs:h-8 sm:h-9 md:h-11 " onSubmit={handleSearch}>
                 <input
                     type="text"
-                    className="peer absolute inset-0 rounded-full text-xs xs:text-sm md:text-base bg-dark-primary text-light-primary outline-none pl-1 xs:pl-3 pr-5 sm:px-5 sm:"
+                    className="absolute inset-0 pl-1 pr-5 text-xs rounded-full outline-none peer xs:text-sm md:text-base bg-dark-primary text-light-primary xs:pl-3 sm:px-5 sm:"
                     placeholder="search for the tools..."
                     value={searchTerm}
                     onChange={(e) => setSearchTerm(e.target.value)}
                 />
-                {filtered.length > 0 && <ul className="hidden peer-focus:block absolute bg-dark-primary max-h-52 overflow-auto top-8 sm:top-14 w-36 xs:w-48 sm:w-72">
+                {filtered.length > 0 && <ul className="absolute hidden overflow-auto peer-focus:block bg-dark-primary max-h-52 top-8 sm:top-14 w-36 xs:w-48 sm:w-72">
                     {filtered.map(({ productName }) => {
                         return (
-                            <li className="py-2 px-5 bg-dark-primary border-b border-light-primary/40 text-light-primary cursor-pointer hover:bg-dark-secondary"
+                            <li className="px-5 py-2 border-b cursor-pointer bg-dark-primary border-light-primary/40 text-light-primary hover:bg-dark-secondary"
                                 key={productName}
                                 onClick={() => handlefilterClick(productName)}>
                                 {productName}
@@ -54,8 +54,8 @@ const SearchBox = () => {
                         )
                     })}
                 </ul>}
-                <button type='submit' className="absolute right-2 top-1/2 -translate-y-1/2 bg-dark-primary text-light-primary font-bold text-lg sm:text-2xl flex justify-center items-center ">
-                    <RiSearchLine className='h-3 xs:h-5 w-3 xs:w-5' />
+                <button type='submit' className="absolute flex items-center justify-center text-lg font-bold -translate-y-1/2 right-2 top-1/2 bg-dark-primary text-light-primary sm:text-2xl ">
+                    <RiSearchLine className='w-3 h-3 xs:h-5 xs:w-5' />
                 </button>
             </form>
         </>
